requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
sqlalchemy_utils==0.41.1
Flask-Babel==4.0.0
Flask-Humanize==0.3.0
humanize==4.9.0
Flask-WTF==1.2.1
